Real Pokemon + Type relationship
Pokemon have changeable stats and have health and stuff, currently no moves change those stats.
The relationship between all types are defined and mostly used. Still need to make sure.
Deleted IPokemon because it was a hassle from the start and I can't use it yet and it's in the way right now.