High Crit Moves
Moves with a high crit rate have a high crit rate.
High Jump Kick and Jump Kick were resolved in another commit and removed from the todo list.
Focus Energy has been moved to priority as it is related to this commit and is naturaly the next thing to work on.